# examples/blog-tools/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "blog-tools",
    "description": "A set of functions for a content management system.",
    "version": "0.1.0",
    "x-dspy-config": {
      "default_model": null,
      "programs_count": 8
    },
    "x-dspy-programs": [
      {
        "name": "TaggerCoT",
        "module_path": "blog_tools.modules.tagger_cot",
        "is_forward_typed": true
      },
      {
        "name": "SummarizerPredict",
        "module_path": "blog_tools.modules.summarizer_predict",
        "is_forward_typed": true
      },
      {
        "name": "HeadlineGeneratorPredict",
        "module_path": "blog_tools.modules.headline_generator_predict",
        "is_forward_typed": true
      },
      {
        "name": "ImageHeadlineGeneratorPredict",
        "module_path": "blog_tools.modules.image_headline_generator_predict",
        "is_forward_typed": true
      },
      {
        "name": "ImageDescriptionGeneratorPredict",
        "module_path": "blog_tools.modules.image_description_generator_predict",
        "is_forward_typed": true
      },
      {
        "name": "TweetExtractorPredict",
        "module_path": "blog_tools.modules.tweet_extractor_predict",
        "is_forward_typed": true
      },
      {
        "name": "SpellCheckerPredict",
        "module_path": "blog_tools.modules.spell_checker_predict",
        "is_forward_typed": true
      },
      {
        "name": "TaggerPredict",
        "module_path": "blog_tools.modules.tagger_predict",
...
  "paths": {
...
    "/ui/{program_name}": {
      "get": {
        "summary": "Program Page",
        "description": "Render the program detail page.",
        "operationId": "program_page_ui__program_name__get",
        "parameters": [
          {
            "name": "program_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Program Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/logs/{program_name}": {
      "get": {
        "summary": "Get Logs",
        "description": "Get recent logs for a program.",
        "operationId": "get_logs_api_logs__program_name__get",
        "parameters": [
          {
            "name": "program_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Program Name"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}